+++ src/features/TodosViewForm/TodosViewForm.component.jsx
@@ -10,7 +10,11 @@ function TodosViewForm({
     <form className="todos-view-form">
       <div className="form-controls">
         <label htmlFor="sort-by">Sort By</label>
-        <select name="sort-by" id="sort-by">
+        <select
+          name="sort-by"
+          id="sort-by"
+          value={sortField}
+        >
           <option value="Title">Title</option>
           <option value="createdTime">Time Created</option>
         </select>
@@ -18,7 +22,7 @@ function TodosViewForm({
         <span className="divider" aria-hidden="true" />
 
         <label htmlFor="direction">Direction</label>
-        <select name="direction" id="direction">
+        <select name="direction" id="direction" value={sortDirection}>
           <option value="asc">Ascending</option>
           <option value="desc">Descending</option>
         </select>